+++ linker/src/main/resources/mybatis/mapper/OrderMapper.xml
@@ -1,0 +1,34 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE mapper
+  PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
+  "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+<mapper namespace="com.linker.ingredient.dao.OrderDao">
+	<!-- 식자재 입고 -->
+	<insert id="addOrder" parameterType="com.linker.ingredient.dto.OrderDto">
+		insert into ingredientorder (ingredientid, orderdate, orderquantity, supplier, orderprice)
+		values (#{ingredientid}, #{orderdate}, #{orderquantity}, #{supplier}, #{orderprice}) on duplicate key
+		update ingredientid = #{ingredientid}
+	</insert>
+	<!-- 식자재 입고내역 수정 -->
+	<update id="changeOrder" parameterType="com.linker.ingredient.dto.OrderDto">
+		update ingredientorder (ingredientid, orderdate, orderquantity, supplier, orderprice)
+		values (#{ingredientid}, #{orderdate}, #{orderquantity}, #{supplier}, #{orderprice})
+		where orderid = #{orderid}
+	</update>
+	<!-- 입고 내역 삭제 -->
+	<delete id="deleteOrder" parameterType="int">
+		delete from ingredientorder where orderid = #{orderid}
+	</delete>
+	<!-- 식자재 입고 내역 개수 -->
+	<select id="orderCount" resultType="int">
+		select count(*) from ingredientorder
+	</select>
+	<!-- 식자재 입고 내역 출력 -->
+	<select id="orderList" parameterType="map" resultType="com.linker.ingredient.dto.OrderDto">
+		select * from ingredientorder
+		inner join ingredient
+		on ingredient.ingredientid = ingredientorder.ingredientid
+		order by orderdate desc
+		limit #{start}, #{count}
+	</select>
+</mapper>
